More internal loads work.
Swapped the internal loads method to work for a single cut at a time, just to get the logic correct.

To get the redundants, and therefore the loads at each segment centroid, we need to do the whole dange thing. So, this will probably have to swap back, but my brain was hurting.